+++ tests/testthat/test-30-add_totals.R
@@ -92,7 +92,7 @@ test_that("do not ignore NA when summing", {
                         "a",  10,
                         "b",  20,
                         "c",  NA
-  )
+                        )
 
   result <-
     px(df) %>%
@@ -101,4 +101,71 @@ test_that("do not ignore NA when summing", {
     dplyr::pull(n)
 
   expect_identical(result, c(NA, 10, 20, NA))
+})
+
+test_that("Add totals work for multiple languages", {
+  # Create px object with multilingual add_totals
+  x <-
+    population_gl %>%
+    px() %>%
+    px_languages(c('en', 'da')) %>%
+    px_elimination(dplyr::tribble(~`variable-code`, ~elimination,
+                                  'gender', 'Total'
+                                  )
+                   ) %>%
+    px_add_totals('gender')
+
+  # px_add_totals adds both 'en' and 'da' level in data. This is a mistake.
+  gender_levels <-
+    x %>%
+    px_data() %>%
+    dplyr::distinct(gender) %>%
+    dplyr::pull(gender)
+
+  expect_identical(gender_levels, c("Total", "male", "female"))
+})
+
+test_that("Add totals adds elimination", {
+  default_elimination <-
+    population_gl %>%
+    px() %>%
+    px_add_totals('gender') %>%
+    px_elimination()
+
+  expect_identical(default_elimination,
+                   dplyr::tibble('variable-code' = 'gender',
+                                 'elimination' = 'Total'
+                                 )
+                   )
+})
+
+test_that("Elimination and values are preserved when changing data frame", {
+  tmp_px_file <- temp_px_file()
+
+  population_gl %>%
+    px() %>%
+    px_add_totals('gender') %>%
+    px_save(tmp_px_file)
+
+  x <-
+    px(tmp_px_file) %>%
+    px_data(population_gl)
+
+  expect_identical(px_elimination(x),
+                   dplyr::tibble('variable-code' = 'gender',
+                                 'elimination' = 'Total'
+                                 )
+                   )
+
+  elimination_values <-
+    x %>%
+    px_values() %>%
+    dplyr::semi_join(px_elimination(x),
+                     by = c('variable-code' = 'variable-code',
+                            'code' = 'elimination'
+                            )
+                     ) %>%
+    dplyr::select("variable-code", "elimination" = "code")
+
+  expect_identical(elimination_values, px_elimination(x))
 })